backend/requirements.txt
# backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pyyaml==6.0.2
requests==2.32.3
matplotlib==3.9.0
networkx==3.3
pandas==2.2.2